+++ stage-marche-devient/stage-marche-devient/Repositories/IReserverRepository.cs
@@ -1,4 +1,4 @@
-﻿namespace stage_marche_devient.Repositorys
+﻿namespace stage_marche_devient.Repositories
 {
     public interface IReserverRepository<T, TForeignKeyOne,TForeingKeyTwo>
     {
